requirements.txt
fastapi
uvicorn[standard]
tensorflow
numpy
pydantic
python-multipart
h5py
requests
pillow
scipy
matplotlib
